+++ changelog.md
@@ -1,5 +1,4 @@
 ------------------------------------------------------
 Version 1.0.0
 ------------------------------------------------------
-- no more deobf jar
-- fixed instant structure items
+- structure placement fix